+++ src/org/pepstock/charba/client/jsinterop/defaults/globals/DefaultPoint.java
@@ -6,7 +6,7 @@ import org.pepstock.charba.client.jsinterop.defaults.IsDefaultPoint;
 public final class DefaultPoint extends DefaultArc implements IsDefaultPoint{
 	
 	/// default radius
-	private static final int DEFAULT_RADIUS = 3;
+	private static final double DEFAULT_RADIUS = 3;
 	// default background color
 	private static final String DEFAULT_BACKGROUND_COLOR = "rgba(0,0,0,0.1)";
 	// default border with
@@ -14,9 +14,9 @@ public final class DefaultPoint extends DefaultArc implements IsDefaultPoint{
 	// default border color
 	private static final String DEFAULT_BORDER_COLOR = "rgba(0,0,0,0.1)";
 	// default hit radius
-	private static final int DEFAULT_HIT_RADIUS = 1;
+	private static final double DEFAULT_HIT_RADIUS = 1;
 	// default hover radius
-	private static final int DEFAULT_HOVER_RADIUS = 4;
+	private static final double DEFAULT_HOVER_RADIUS = 4;
 	// default hover border width
 	private static final int DEFAULT_HOVER_BORDER_WIDTH = 1;
 
@@ -48,7 +48,7 @@ public final class DefaultPoint extends DefaultArc implements IsDefaultPoint{
 	 * @see org.pepstock.charba.client.jsinterop.options.elements.point.IsDefaultPoint#getRadius()
 	 */
 	@Override
-	public int getRadius() {
+	public double getRadius() {
 		return DEFAULT_RADIUS;
 	}
 
@@ -64,7 +64,7 @@ public final class DefaultPoint extends DefaultArc implements IsDefaultPoint{
 	 * @see org.pepstock.charba.client.jsinterop.options.elements.point.IsDefaultPoint#getHitRadius()
 	 */
 	@Override
-	public int getHitRadius() {
+	public double getHitRadius() {
 		return DEFAULT_HIT_RADIUS;
 	}
 
@@ -72,7 +72,7 @@ public final class DefaultPoint extends DefaultArc implements IsDefaultPoint{
 	 * @see org.pepstock.charba.client.jsinterop.options.elements.point.IsDefaultPoint#getHoverRadius()
 	 */
 	@Override
-	public int getHoverRadius() {
+	public double getHoverRadius() {
 		return DEFAULT_HOVER_RADIUS;
 	}
